Use new ProsopoApiError and move dev setup stuff to tests. Create tests tsconfig.json
