+++ src/errors.ts
@@ -11,11 +11,17 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import { ProsopoEnvError } from '@prosopo/common'
+import { ProsopoApiError } from '@prosopo/common'
 
-export const handleErrors = (err: ProsopoEnvError, req, res, next) => {
-    return res.status(500).json({
-        message: err.message,
+export const handleErrors = (err: ProsopoApiError, req, res, next) => {
+    let message = err.message
+    try {
+        message = JSON.parse(err.message)
+    } catch {
+        console.debug('Invalid JSON error message')
+    }
+    return res.status(err.code).json({
+        message,
         name: err.name,
     })
 }
