+++ src/tasks/index.ts
@@ -12,4 +12,3 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 export * from './tasks'
-export * from './setup'
